--- webpack/ForemanYupana/Components/UploadsDashboard/index.js
@@ -1,1 +1,0 @@
-export { default } from './UploadsDashboard';
