+++ webpack/ForemanYupana/ForemanYupana.js
@@ -1,12 +1,12 @@
 import React from 'react';
-import UploadsDashboard from './Components/UploadsDashboard';
+import Dashboard from './Components/Dashboard';
 import {
   generating,
   uploading,
-} from './Components/UploadsDashboard/UploadsDashboard.fixtures';
+} from './Components/Dashboard/Dashboard.fixtures';
 
 const ForemanYupana = () => (
-  <UploadsDashboard generating={generating} uploading={uploading} />
+  <Dashboard generating={generating} uploading={uploading} />
 );
 
 export default ForemanYupana;
